- synced with SVN

+++ DependentExtensions/IrrlichtDemo/CDemo.cpp
@@ -5,7 +5,7 @@
  * This file was taken from RakNet 4.082.
  * Please see licenses/RakNet license.txt for the underlying license and related copyright.
  *
- * Modified work: Copyright (c) 2016-2017, SLikeSoft UG (haftungsbeschränkt)
+ * Modified work: Copyright (c) 2016-2018, SLikeSoft UG (haftungsbeschränkt)
  *
  * This source code was modified by SLikeSoft. Modifications are licensed under the MIT-style
  * license found in the license.txt file in the root directory of this source tree.
@@ -1167,7 +1167,7 @@ void CDemo::UpdateRakNet(void)
 			if (packet->guid!=rakPeer->GetGuidFromSystemAddress(SLNet::UNASSIGNED_SYSTEM_ADDRESS))
 			{
 				char hostIP[32];
-				packet->systemAddress.ToString(false,hostIP,32);
+				packet->systemAddress.ToString(false,hostIP,static_cast<size_t>(32));
 				SLNET_VERIFY(rakPeer->Connect(hostIP,packet->systemAddress.GetPort(),0,0) == SLNet::CONNECTION_ATTEMPT_STARTED);
 			}
 			break;
